feat: re-organize fee tx execution into an alarm
