+++ public/manifest.json
@@ -4,7 +4,14 @@
   "version": "0.4.3",
   "author": "th8ta",
   "description": "Secure wallet management for Arweave",
-  "permissions": ["activeTab", "tabs", "storage", "contextMenus", "<all_urls>"],
+  "permissions": [
+    "activeTab",
+    "tabs",
+    "storage",
+    "contextMenus",
+    "<all_urls>",
+    "alarms"
+  ],
   "background": {
     "scripts": ["build/scripts/background.js"]
   },
